refactor: relocate ExceptionHandlerEndpointDefinition.cs
